app/requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.2
certifi==2020.4.5.1
Django==3.0.5
django-crispy-forms==1.9.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
sqlparse==0.3.1
wrapt==1.11.2
